Helper for checking media query through React
