+++ src/components/client/pages/home/helpers.jsx
@@ -1,0 +1,19 @@
+import { useMediaQuery } from '@mui/material';
+
+function CheckMediaQuery() {
+  const matchesMobile = useMediaQuery('(min-width:20em)');
+  const matchesTablet = useMediaQuery('(min-width:30em)');
+  const matchesDesktop = useMediaQuery('(min-width:48em)');
+
+  if (matchesDesktop) {
+    return 4;
+  }
+  if (matchesTablet) {
+    return 2;
+  }
+  if (matchesMobile) {
+    return 1;
+  }
+}
+
+export { CheckMediaQuery };
